Revamped the system in which Piece::ValidSpots gives locations.
It now puts the capture command within a special captureMap within each piece, and that capture map is used to describe the behavior of the piece when it captures.

With this new system we can have pieces destroy other pieces, move, and spawn all in one movement with complete flexibility.