Game now has a Multi-click mode.
When you right click on an eligable piece (The queen in this example) you are presented with a purple cursor.

You are allowed to click x number of times (defined within the piece you right clicked on)

After x times, it will execute the command within the processClicks function of the selected piece.

In this example, right clicking on a queen and clicking on another piece will bring the queen to that piece's location.